Revert everything after map TS migration
Somehow the TS migration broke two things:
 - getFeatureListBounds() for point feature lists returns the same points for both corners
 - The "Add Point" button on the map context menu stopped working.

Revert it for now and do the migration later.
