+++ package.json
@@ -70,7 +70,6 @@
   "devDependencies": {
     "@types/json-logic-js": "^2.0.4",
     "@types/leaflet": "^1.9.0",
-    "@types/leaflet.locatecontrol": "^0.74.5",
     "gh-pages": "^4.0.0",
     "prettier": "3.3.0"
   }
