+++ src/geojson-schema.js
